services/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
scikit-learn==1.3.1
numpy==1.26.4
catboost==1.2.2
requests==2.32.3
boto3
load-dotenv
pyarrow
fastparquet